Colorize the output (#66)
Adds some colors to the logs and also some empty space around the log
lines from try-runtime-cli to make it more clear what belongs to what.
Before I always had the problem that there was a wall of text and I
didn't know when something started and something ended. This pr makes it
a little bit better. The ashetics are discussable and clearly can be
improved.

---------

Co-authored-by: Liam Aharon <liam.aharon@hotmail.com>